// apps/api/src/generated/kysely/valuations/AssetType.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Represents the enum valuations.AssetType */
enum AssetType {
  CONVERTIBLE = 'CONVERTIBLE',
  CURRENCY = 'CURRENCY',
  EMPLOYEE_STOCK_OPTIONS = 'EMPLOYEE_STOCK_OPTIONS',
  EQUITY = 'EQUITY',
  EQUITY_UNKNOWN_SHARES = 'EQUITY_UNKNOWN_SHARES',
  LP_INTEREST_POINT = 'LP_INTEREST_POINT',
  SPV_INTEREST_POINT = 'SPV_INTEREST_POINT',
  UNKNOWN = 'UNKNOWN',
  FUND_OUTSTANDING_COMMITMENT = 'FUND_OUTSTANDING_COMMITMENT',
  ACCRUED_INCOME = 'ACCRUED_INCOME',
};

export default AssetType;